move jvmId/filename selectors out of tab
