+++ src/app/RecordingAnalytics/RecordingAnalytics.tsx
@@ -15,10 +15,26 @@
  */
 
 import { BreadcrumbPage } from '@app/BreadcrumbPage/BreadcrumbPage';
+import { modalPrefillClearIntent, RootState } from '@app/Shared/Redux/ReduxStore';
+import { NotificationCategory, RecordingDirectory } from '@app/Shared/Services/api.types';
+import { ServiceContext } from '@app/Shared/Services/Services';
+import { useSubscriptions } from '@app/utils/hooks/useSubscriptions';
 import { getActiveTab, switchTab } from '@app/utils/utils';
 import { useCryostatTranslation } from '@i18n/i18nextUtil';
-import { Card, CardBody, Tab, Tabs, TabTitleText } from '@patternfly/react-core';
+import {
+  Card,
+  CardBody,
+  Tab,
+  Tabs,
+  TabTitleText,
+  Toolbar,
+  ToolbarContent,
+  ToolbarGroup,
+  ToolbarItem,
+} from '@patternfly/react-core';
+import { SimpleDropdown, SimpleDropdownItem } from '@patternfly/react-templates';
 import * as React from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router-dom-v5-compat';
 import { Queries } from './queries/Queries';
 
@@ -28,8 +44,139 @@ enum RecordingAnalyticsTab {
 
 export const RecordingAnalytics: React.FC = () => {
   const { t } = useCryostatTranslation();
-  const { search, pathname } = useLocation();
+  const context = React.useContext(ServiceContext);
+  const addSubscription = useSubscriptions();
+  const location = useLocation();
   const navigate = useNavigate();
+  const dispatch = useDispatch();
+  const modalPrefill = useSelector((state: RootState) => state.modalPrefill);
+  const { search, pathname } = location;
+
+  const [jvmId, setJvmId] = React.useState('');
+  const [recordingDirectories, setRecordingDirectories] = React.useState([] as RecordingDirectory[]);
+  const [filename, setFilename] = React.useState('');
+
+  const refreshRecordingDirectories = React.useCallback(() => {
+    addSubscription(
+      context.api.doGet<RecordingDirectory[]>('fs/recordings', 'beta').subscribe((v) => {
+        setRecordingDirectories(v);
+      }),
+    );
+  }, [addSubscription, context.api]);
+
+  React.useEffect(() => {
+    refreshRecordingDirectories();
+  }, [refreshRecordingDirectories]);
+
+  React.useEffect(() => {
+    addSubscription(
+      context.notificationChannel.messages(NotificationCategory.ArchivedRecordingCreated).subscribe(() => {
+        refreshRecordingDirectories();
+      }),
+    );
+  }, [addSubscription, context.notificationChannel, refreshRecordingDirectories]);
+
+  React.useEffect(() => {
+    addSubscription(
+      context.notificationChannel.messages(NotificationCategory.ArchivedRecordingDeleted).subscribe(() => {
+        refreshRecordingDirectories();
+      }),
+    );
+  }, [addSubscription, context.notificationChannel, refreshRecordingDirectories]);
+
+  React.useEffect(() => {
+    const stateData = location.state as Record<string, unknown> | null;
+    const reduxData = modalPrefill.route === location.pathname ? (modalPrefill.data as Record<string, unknown>) : null;
+
+    const prefillJvmId = (stateData?.jvmId || reduxData?.jvmId) as string | undefined;
+    const prefillFilename = (stateData?.filename || reduxData?.filename) as string | undefined;
+
+    if (prefillJvmId && recordingDirectories.some((d) => d.jvmId === prefillJvmId)) {
+      setJvmId(prefillJvmId);
+
+      if (prefillFilename) {
+        const directory = recordingDirectories.find((d) => d.jvmId === prefillJvmId);
+        if (directory && directory.recordings.some((r) => r.name === prefillFilename)) {
+          setFilename(prefillFilename);
+        }
+      }
+
+      dispatch(modalPrefillClearIntent());
+      if (location.state) {
+        navigate(`${location.pathname}${location.search}${location.hash}`, { replace: true, state: null });
+      }
+    }
+  }, [
+    recordingDirectories,
+    location.state,
+    location.pathname,
+    location.search,
+    location.hash,
+    modalPrefill,
+    dispatch,
+    navigate,
+  ]);
+
+  const jvmIds = React.useMemo(() => recordingDirectories.map((e) => e.jvmId), [recordingDirectories]);
+
+  const filenames = React.useMemo(() => {
+    const directory = recordingDirectories.find((d) => d.jvmId === jvmId);
+    return directory ? directory.recordings.map((r) => r.name) : [];
+  }, [recordingDirectories, jvmId]);
+
+  const jvmIdItems = React.useMemo(() => {
+    const a: SimpleDropdownItem[] = jvmIds
+      .map(
+        (id) =>
+          ({
+            value: id,
+            onClick: () => {
+              setJvmId(id);
+              setFilename('');
+            },
+            content: id,
+          }) as SimpleDropdownItem,
+      )
+      .concat([
+        {
+          value: '',
+          isDivider: true,
+        },
+        {
+          value: 'Clear Selection',
+          onClick: () => {
+            setJvmId('');
+            setFilename('');
+          },
+          content: 'Clear Selection',
+        },
+      ]);
+    return a;
+  }, [jvmIds]);
+
+  const filenameItems = React.useMemo(() => {
+    const a: SimpleDropdownItem[] = filenames
+      .map(
+        (f) =>
+          ({
+            value: f,
+            onClick: () => setFilename(f),
+            content: f,
+          }) as SimpleDropdownItem,
+      )
+      .concat([
+        {
+          value: '',
+          isDivider: true,
+        },
+        {
+          value: 'Clear Selection',
+          onClick: () => setFilename(''),
+          content: 'Clear Selection',
+        },
+      ]);
+    return a;
+  }, [filenames]);
 
   const activeTab = React.useMemo(
     () => getActiveTab(search, 'tab', Object.values(RecordingAnalyticsTab), RecordingAnalyticsTab.QUERIES),
@@ -46,12 +193,28 @@ export const RecordingAnalytics: React.FC = () => {
     <BreadcrumbPage pageTitle="Analytics">
       <Card isFullHeight>
         <CardBody isFilled>
+          <Toolbar>
+            <ToolbarContent>
+              <ToolbarGroup>
+                <ToolbarItem>
+                  <SimpleDropdown toggleContent={jvmId || 'JVM ID'} initialItems={jvmIdItems} />
+                </ToolbarItem>
+                <ToolbarItem>
+                  <SimpleDropdown
+                    toggleContent={filename || 'Filename'}
+                    isDisabled={!jvmId}
+                    initialItems={filenameItems}
+                  />
+                </ToolbarItem>
+              </ToolbarGroup>
+            </ToolbarContent>
+          </Toolbar>
           <Tabs activeKey={activeTab} onSelect={onTabSelect}>
             <Tab
               eventKey={RecordingAnalyticsTab.QUERIES}
               title={<TabTitleText>{t('RecordingAnalytics.QUERIES_TAB_TITLE')}</TabTitleText>}
             >
-              <Queries />
+              <Queries jvmId={jvmId} filename={filename} />
             </Tab>
           </Tabs>
         </CardBody>
